* move listeners and their implementations to seperate classes / files * make sure the normal bindViewHolder call is not called if legacyBindViewMode is activated (to prevent twice the binds) ^^   * this is anyways just a hack for the weird case of a library not properly implementing the `onBindViewHolder` method (in case we depend on the event being forwarded to us..) * add FastAdapter to the `OnCreateViewHolderListener` methods so it is properly possible to get the handled TypeInstance from the viewType and it's parent * add getEventHooks() method
