+++ library-core/src/main/java/com/mikepenz/fastadapter/IClickable.java
@@ -1,14 +1,16 @@
 package com.mikepenz.fastadapter;
 
+import com.mikepenz.fastadapter.listeners.OnClickListener;
+
 /**
  * Created by mikepenz on 25.02.16.
  */
 public interface IClickable<Item extends IItem> {
-    Item withOnItemPreClickListener(FastAdapter.OnClickListener<Item> onItemPreClickListener);
+    Item withOnItemPreClickListener(OnClickListener<Item> onItemPreClickListener);
 
-    FastAdapter.OnClickListener<Item> getOnPreItemClickListener();
+    OnClickListener<Item> getOnPreItemClickListener();
 
-    Item withOnItemClickListener(FastAdapter.OnClickListener<Item> onItemClickListener);
+    Item withOnItemClickListener(OnClickListener<Item> onItemClickListener);
 
-    FastAdapter.OnClickListener<Item> getOnItemClickListener();
+    OnClickListener<Item> getOnItemClickListener();
 }
